sonnow_back/users/views/user_view.py

from django.utils.decorators import method_decorator
from django.views.decorators.csrf import csrf_protect
from rest_framework import status
from rest_framework.response import Response
from rest_framework.views import APIView

class UserView(APIView):

    @method_decorator(csrf_protect)
    def get(self, request):
        if not request.user.is_authenticated:
            return Response(status=status.HTTP_401_UNAUTHORIZED)
        user = request.user
        return Response({
            "username": user.username,
            "email": user.email,
        })


