sonnow_back/requirements.txt
djangorestframework-simplejwt~=5.5.0
Django~=4.2.20
djangorestframework~=3.15.2